TOFIXUP: Move config into the manifest.yml file.
We're able to make use of cf v7 variable substiution[1] to move some of
our config out of concourse CLI commands and into the manifest.yml.

This has the advantage of letting cloud foundry work out the correct
order of steps (eg scaling, mapping routes etc).

We can have a separate vars file for each environment and then
interpolate that through a concoruse env var.

Whilst the change here is modest, it is a proof of concept that will
make the case for a bigger shift in our other Accounts apps. Those apps
have far more environment variables and `cf` commands in the concourse
pipeline which we should be able to clear up with this method

[1]: https://docs.cloudfoundry.org/devguide/deploy-apps/manifest-attributes.html#variable-substitution